requirements.txt
streamlit
pandas
numpy
torch
joblib
pvlib
matplotlib
plotly
scikit-learn
statsmodels
requests
urllib3
python-dotenv